split actions into separate libs
Move the "describe a bug" action to a separate "actions.describe"
library.

This will make room for us to implement additional actions.

Generalize the main helga_bugzilla() method so that each action passes a
tuple of callback methods to call, rather than hard-coding the list of
specific callback methods. This allows me to define completely separate
callbacks for each action.

Move the "describe_bugs()" method to a util lib, so this can be shared
across actions.