add external tracker searching
Use txbugzilla 1.2.0's new ability to search for BZs according to
external trackers.

Add some common external trackers that we use in the Ceph team.

Add custom HelgaBugzilla exceptions for some cases. This lets us bubble up
the error messages all the way to the Twisted errback, so helga will
print the messages in the chat.